Cleaned op duplicate conditions code
