+++ frontend/projects/valtimo/case/src/lib/models/case-widget.model.ts
@@ -23,6 +23,7 @@ import {
   WidgetTableContent,
 } from './case-widget-content.model';
 import {CaseWidgetDisplayType} from './case-widget-display.model';
+import {Condition} from '@valtimo/shared';
 
 enum CaseWidgetType {
   FIELDS = 'fields',
@@ -160,21 +161,6 @@ interface CustomCaseWidgetConfig {
   [componentKey: string]: Type<any>;
 }
 
-enum Operator {
-  NOT_EQUAL_TO = '!=',
-  EQUAL_TO = '==',
-  GREATER_THAN = '>',
-  GREATER_THAN_OR_EQUAL_TO = '>=',
-  LESS_THAN = '<',
-  LESS_THAN_OR_EQUAL_TO = '<=',
-}
-
-interface Condition {
-  path: string;
-  operator: string;
-  value: string;
-}
-
 export {
   BasicCaseWidget,
   CaseWidget,
@@ -200,6 +186,4 @@ export {
   CaseWidgetPackResultItemsByRow,
   FormioCaseWidgetWidgetWithUuid,
   MaxRectsResult,
-  Operator,
-  Condition,
 };
