+++ frontend/projects/valtimo/dashboard/src/lib/data-sources/case-count/case-count.specification.ts
@@ -26,12 +26,6 @@ export const caseCountDataSourceSpecification: DataSourceSpecification = {
       title: 'Fallzahl',
       documentDefinition: 'Falltyp (erforderlich)',
       documentDefinitionHelperText: 'Der Falltyp, für den die Anzahl abgerufen wird',
-      '!=': 'Nicht gleichzusetzen mit',
-      '==': 'Gleich',
-      '>': 'Größer als',
-      '>=': 'Größer als oder gleich wie',
-      '<': 'Weniger als',
-      '<=': 'Gleich oder kleiner als',
       path: 'Pfad',
       operator: 'Operator',
       value: 'Wert',
@@ -43,12 +37,6 @@ export const caseCountDataSourceSpecification: DataSourceSpecification = {
       title: 'Case count',
       documentDefinition: 'Case type (required)',
       documentDefinitionHelperText: 'The case type for which the count is retrieved',
-      '!=': 'Not equal to',
-      '==': 'Equal to',
-      '>': 'Greater than',
-      '>=': 'Greater than or equal to',
-      '<': 'Less than',
-      '<=': 'Less than or equal to',
       path: 'Path (required)',
       operator: 'Operator',
       value: 'Value',
@@ -60,12 +48,6 @@ export const caseCountDataSourceSpecification: DataSourceSpecification = {
       title: 'Aantal dossiers',
       documentDefinition: 'Dossiertype (vereist)',
       documentDefinitionHelperText: 'Het dossiertype waarvoor de telling wordt opgehaald',
-      '!=': 'Niet gelijk aan',
-      '==': 'Gelijk aan',
-      '>': 'Groter dan',
-      '>=': 'Groter dan of gelijk aan',
-      '<': 'Minder dan',
-      '<=': 'Minder dan of gelijk aan',
       path: 'Pad',
       operator: 'Operator',
       value: 'Waarde',
