+++ frontend/projects/valtimo/dashboard/src/lib/data-sources/case-counts/case-counts.specification.ts
@@ -26,12 +26,6 @@ export const caseCountsDataSourceSpecification: DataSourceSpecification = {
       title: 'Mehrere Fallanzahllen',
       documentDefinition: 'Falltyp (erforderlich)',
       documentDefinitionHelperText: 'Der Falltyp, für den die Anzahllen abgeruft wirden',
-      '!=': 'Nicht gleichzusetzen mit',
-      '==': 'Gleich',
-      '>': 'Größer als',
-      '>=': 'Größer als oder gleich wie',
-      '<': 'Weniger als',
-      '<=': 'Gleich oder kleiner als',
       path: 'Pfad',
       operator: 'Operator',
       value: 'Wert',
@@ -44,12 +38,6 @@ export const caseCountsDataSourceSpecification: DataSourceSpecification = {
       title: 'Multiple case counts',
       documentDefinition: 'Case type (required)',
       documentDefinitionHelperText: 'The case type for which the counts are retrieved',
-      '!=': 'Not equal to',
-      '==': 'Equal to',
-      '>': 'Greater than',
-      '>=': 'Greater than or equal to',
-      '<': 'Less than',
-      '<=': 'Less than or equal to',
       path: 'Path (required)',
       operator: 'Operator',
       value: 'Value',
@@ -62,12 +50,6 @@ export const caseCountsDataSourceSpecification: DataSourceSpecification = {
       title: 'Meerdere dossieraantallen',
       documentDefinition: 'Dossiertype (vereist)',
       documentDefinitionHelperText: 'Het dossiertype waarvoor de aantallen worden opgehaald',
-      '!=': 'Niet gelijk aan',
-      '==': 'Gelijk aan',
-      '>': 'Groter dan',
-      '>=': 'Groter dan of gelijk aan',
-      '<': 'Minder dan',
-      '<=': 'Minder dan of gelijk aan',
       path: 'Pad',
       operator: 'Operator',
       value: 'Waarde',
