+++ frontend/projects/valtimo/dashboard/src/lib/data-sources/case-group-by/case-group-by.specification.ts
@@ -26,12 +26,6 @@ export const caseGroupByDataSourceSpecification: DataSourceSpecification = {
       title: 'Gruppe',
       documentDefinition: 'Falltyp (erforderlich)',
       documentDefinitionHelperText: 'Der Falltyp, für den die Anzahl abgerufen wird',
-      '!=': 'Nicht gleichzusetzen mit',
-      '==': 'Gleich',
-      '>': 'Größer als',
-      '>=': 'Größer als oder gleich wie',
-      '<': 'Weniger als',
-      '<=': 'Gleich oder kleiner als',
       path: 'Pfad',
       pathHelperText: 'Der Pfad innerhalb der Fall, deren Werte gruppiert werden sollen',
       operator: 'Operator',
@@ -49,12 +43,6 @@ export const caseGroupByDataSourceSpecification: DataSourceSpecification = {
       title: 'Group by',
       documentDefinition: 'Case type (required)',
       documentDefinitionHelperText: 'The case type for which the count is retrieved',
-      '!=': 'Not equal to',
-      '==': 'Equal to',
-      '>': 'Greater than',
-      '>=': 'Greater than or equal to',
-      '<': 'Less than',
-      '<=': 'Less than or equal to',
       path: 'Path (required)',
       pathHelperText: 'The path within the case of which the values should be grouped',
       operator: 'Operator',
@@ -72,12 +60,6 @@ export const caseGroupByDataSourceSpecification: DataSourceSpecification = {
       title: 'Groepering',
       documentDefinition: 'Dossiertype (vereist)',
       documentDefinitionHelperText: 'Het dossiertype waarvoor de telling wordt opgehaald',
-      '!=': 'Niet gelijk aan',
-      '==': 'Gelijk aan',
-      '>': 'Groter dan',
-      '>=': 'Groter dan of gelijk aan',
-      '<': 'Minder dan',
-      '<=': 'Minder dan of gelijk aan',
       path: 'Pad',
       pathHelperText: 'Het pad binnen het dossier waarvan de waardes gegroepeerd moeten worden',
       operator: 'Operator',
