+++ frontend/projects/valtimo/dashboard/src/lib/data-sources/case-group-by/components/case-group-by-configuration/case-group-by-configuration.component.ts
@@ -26,7 +26,6 @@ import {
 import {
   ConfigurationOutput,
   DataSourceConfigurationComponent,
-  Operator,
   QueryCondition,
 } from '../../../../models';
 import {BehaviorSubject, combineLatest, map, Observable, startWith, Subscription} from 'rxjs';
@@ -46,8 +45,8 @@ import {DocumentService} from '@valtimo/document';
 import {ListItem} from 'carbon-components-angular';
 import {ListItemWithId, MultiInputValues, ValuePathSelectorPrefix} from '@valtimo/components';
 import {TranslateService} from '@ngx-translate/core';
-import {WidgetTranslationService} from '../../../../services';
 import {isEqual} from 'lodash';
+import {Operator} from '@valtimo/shared';
 
 @Component({
   standalone: false,
@@ -128,7 +127,7 @@ export class CaseGroupByConfigurationComponent
       map(() =>
         this._OPERATORS.map(operator => ({
           id: operator,
-          content: this.widgetTranslationService.instant(operator, this.dataSourceKey),
+          content: this.translateService.instant('condition.operator.' + operator),
           selected: false,
         }))
       )
@@ -161,8 +160,7 @@ export class CaseGroupByConfigurationComponent
   constructor(
     private readonly fb: FormBuilder,
     private readonly documentService: DocumentService,
-    private readonly translateService: TranslateService,
-    private readonly widgetTranslationService: WidgetTranslationService
+    private readonly translateService: TranslateService
   ) {}
 
   public ngOnInit(): void {
