+++ frontend/projects/valtimo/dashboard/src/lib/models/data-source.model.ts
@@ -28,19 +28,10 @@ interface DataSourceSpecification {
   };
 }
 
-enum Operator {
-  NOT_EQUAL_TO = '!=',
-  EQUAL_TO = '==',
-  GREATER_THAN = '>',
-  GREATER_THAN_OR_EQUAL_TO = '>=',
-  LESS_THAN = '<',
-  LESS_THAN_OR_EQUAL_TO = '<=',
-}
-
 interface QueryCondition {
   queryPath: string;
   queryOperator: string;
   queryValue: string;
 }
 
-export {DataSourceSpecification, Operator, QueryCondition};
+export {DataSourceSpecification, QueryCondition};
