+++ frontend/projects/valtimo/shared/src/lib/models/index.ts
@@ -37,3 +37,4 @@ export * from './global-configuration.model';
 export * from './case-management.model';
 export * from './display-type.model';
 export * from './modal.model';
+export * from './condition.model';
